feat(prussia_rt): Move v_common exc handling to own file
